Cleanup CeloBackend
There is an almost identical implementation at contracts/celo_backend.go
and only that one is actually used.

Move GetExchangeRates out of CeloBackend

The CeloBackend is meant to provide access to contract calling and
execution, not to have any business logic itself.

This change also allows the function to be called with arbitrary
ContractCallers instead of being needlessly tied to a CeloBackend.

Move GetBalanceERC20 out of CeloBackend

The CeloBackend is meant to provide access to contract calling and
execution, not to have any business logic itself.

This change also allows the function to be called with arbitrary
ContractCallers instead of being needlessly tied to a CeloBackend.

Move GetFeeBalance out of CeloBackend

The CeloBackend is meant to provide access to contract calling and
execution, not to have any business logic itself.
