+++ core/txpool/legacypool/celo_legacypool.go
@@ -24,7 +24,7 @@ func (pool *LegacyPool) filter(list *list, addr common.Address, gasLimit uint64)
 func (pool *LegacyPool) getBalances(address common.Address, currencies []common.Address) map[common.Address]*uint256.Int {
 	balances := make(map[common.Address]*uint256.Int, len(currencies))
 	for _, curr := range currencies {
-		balances[curr] = uint256.MustFromBig(pool.celoBackend.GetFeeBalance(address, &curr))
+		balances[curr] = uint256.MustFromBig(contracts.GetFeeBalance(pool.celoBackend, address, &curr))
 	}
 	return balances
 }
@@ -34,7 +34,7 @@ func (pool *LegacyPool) recreateCeloProperties() {
 		ChainConfig: pool.chainconfig,
 		State:       pool.currentState,
 	}
-	currentRates, err := pool.celoBackend.GetExchangeRates()
+	currentRates, err := contracts.GetExchangeRates(pool.celoBackend)
 	if err != nil {
 		log.Error("Error trying to get exchange rates in txpool.", "cause", err)
 	}
